[SDL Widget] Move 'state' property and state management from button class onto widget base class.
Also move widget function definitions to separate implementation file instead of continuing to clutter the header file.
Current issue: Button state is incorrectly allowed to go back to Hover when it should be Off.
Button should appear as Hover when the mouse drifts off of it after clicking (currently stays active, but no callback will fire if the mouse is released off of the button).